fix: 动态列表、个人主主页为发送成功的评论增加删除

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/dynamic/list/DynamicPresenter.java
@@ -339,14 +339,19 @@ public class DynamicPresenter extends AppBasePresenter<DynamicContract.View>
 
     @Override
     public void deleteCommentV2(DynamicDetailBeanV2 dynamicBean, int dynamicPositon, long comment_id, int commentPosition) {
-        mRootView.getListDatas().get(dynamicPositon).setFeed_comment_count(dynamicBean.getFeed_comment_count() - 1);
+        if(comment_id>0) {
+            mRootView.getListDatas().get(dynamicPositon).setFeed_comment_count(dynamicBean.getFeed_comment_count() - 1);
+        }
+
         mDynamicDetailBeanV2GreenDao.insertOrReplace(mRootView.getListDatas().get(dynamicPositon));
-        if (!dynamicBean.getComments().isEmpty()){
+        if (!dynamicBean.getComments().isEmpty()) {
             mDynamicCommentBeanGreenDao.deleteSingleCache(dynamicBean.getComments().get(commentPosition));
             mRootView.getListDatas().get(dynamicPositon).getComments().remove(commentPosition);
         }
         mRootView.refreshData(dynamicPositon);
-        mDynamicRepository.deleteCommentV2(dynamicBean.getId(), comment_id);
+        if (comment_id > 0) {
+            mDynamicRepository.deleteCommentV2(dynamicBean.getId(), comment_id);
+        }
     }
 
     @Override
